# app/backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: QuickEx Backend API
  description: >
    QuickEx is a Stellar-based payment platform enabling payment links,
    recurring subscriptions, username resolution, marketplace trading,
    fiat on/off ramps, and webhook-driven event notifications.

    **Base URL (production):** `https://api.quickex.example.com`
    **Base URL (development):** `http://localhost:3000`

    Interactive Swagger UI is available at `/docs` when the server is running.
  version: 1.0.0
  contact:
    name: QuickEx API Support
    url: https://github.com/quickex/quickex
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.quickex.example.com
    description: Production
...
security:
  - ApiKeyAuth: []
  - {}
paths:
...
  /notifications/preferences/{publicKey}:
    get:
      tags: [Notifications]
      operationId: listPreferences
      summary: List notification preferences for a wallet
      parameters:
        - name: publicKey
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of preferences
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NotificationPreference"

    put:
      tags: [Notifications]
      operationId: upsertPreference
      summary: Create or update a notification channel preference
      parameters:
        - name: publicKey
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertNotificationPreference"
      responses:
        "200":
          description: Preference updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationPreference"
  /notifications/preferences:
    get:
      tags: [Notifications]
      operationId: getCurrentPreferences
      summary: Get notification preferences for the authenticated user
      responses:
        "200":
          description: Current notification preferences
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NotificationPreference"

    patch:
      tags: [Notifications]
      operationId: updateCurrentPreferences
      summary: Update notification preferences for the authenticated user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNotificationPreferences"
      responses:
        "200":
          description: Updated notification preferences
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NotificationPreference"
        "400":
          description: At least one channel must remain enabled
  /notifications/preferences/{publicKey}/{channel}:
    delete:
      tags: [Notifications]
      operationId: disableChannel
      summary: Opt-out of a notification channel
      parameters:
        - name: publicKey
          in: path
          required: true
          schema:
            type: string
        - name: channel
          in: path
          required: true
          schema:
            type: string
            enum: [email, telegram, push, webhook]
      responses:
        "204":
          description: Channel disabled

  # ---------------------------------------------------------------------------
  # Marketplace
  # ---------------------------------------------------------------------------
...
components:
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Optional API key for higher rate limits

  # -------------------------------------------------------------------------
  # Parameters
  # -------------------------------------------------------------------------
...
  schemas:
...
    NotificationPreference:
      type: object
      properties:
        id:
          type: string
        publicKey:
          type: string
        channel:
          type: string
          enum: [email, telegram, push, webhook]
        email:
          type: string
        pushToken:
          type: string
        webhookUrl:
          type: string
        webhookSecret:
          type: string
        events:
          type: array
          items:
            type: string
          nullable: true
        minAmountStroops:
          type: string
        enabled:
          type: boolean
    UpsertNotificationPreference:
      type: object
      required: [channel, enabled]
      properties:
        channel:
          type: string
          enum: [email, telegram, push, webhook]
        email:
          type: string
        pushToken:
          type: string
        webhookUrl:
          type: string
        webhookSecret:
          type: string
        events:
          type: array
          items:
            type: string
          nullable: true
        minAmountStroops:
          type: number
        enabled:
          type: boolean
    UpdateNotificationPreferences:
      type: object
      required: [preferences]
      properties:
        preferences:
          type: array
          items:
            $ref: "#/components/schemas/UpsertNotificationPreference"

    # --- Marketplace ---
...